+++ src/main/java/vn/kms/ngaythobet/domain/validation/NotExpiredTime.java
@@ -1,0 +1,36 @@
+// Copyright (c) 2015 KMS Technology, Inc.
+package vn.kms.ngaythobet.domain.validation;
+
+import static java.lang.annotation.ElementType.ANNOTATION_TYPE;
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import javax.validation.Constraint;
+import javax.validation.Payload;
+
+import vn.kms.ngaythobet.domain.core.AuditableEntity;
+
+@Target({TYPE, ANNOTATION_TYPE})
+@Retention(RUNTIME)
+@Constraint(validatedBy = NotExpiredTimeValidator.class)
+@Documented
+public @interface NotExpiredTime {
+
+    String message() default "{validation.expiredTime.message}";
+
+    Class<?>[] groups() default {};
+
+    Class<? extends Payload>[] payload() default {};
+    
+    Class<? extends AuditableEntity> type();
+
+    String entityId();
+
+    String targetField();
+    
+    String fieldName();
+}